Moved and renamed more stuff.
