+++ halo/upgrades/map.cpp
